+++ packages/jivs-engine/src/Interfaces/InputValidator.ts
@@ -223,12 +223,13 @@ export interface InputValidatorDescriptor {
      */
     summaryMessagel10n?: string | null | undefined;
 
-    /**
-     * Handy way to allow users to enter known properties without getting ts errors.
-     * However, they can improve things if they typecast to the appropriate
-     * inputValidator's Descriptor.
-     */
-    [propName: string]: any;    
+    //!! Interferes with intellisense support for building with known properties    
+    // /**
+    //  * Handy way to allow users to enter known properties without getting ts errors.
+    //  * However, they can improve things if they typecast to the appropriate
+    //  * inputValidator's Descriptor.
+    //  */
+    // [propName: string]: any;    
 }
 
 /**